+++ stage3/xfce/stage3-xfce.sh
@@ -601,7 +601,7 @@ pushd lightdm-1.32.0
 patch -Np1 -i ../../patches/lightdm-1.32.0-xsession.patch
 patch -Np1 -i ../../patches/lightdm-1.32.0-fixmemoryleak.patch
 sed -i 's|initdir = ${sysconfdir}/init|initdir = /tmp/.mbs_trash/init|' data/Makefile.in
-echo 'u lightdm 430 "LightDM Daemon" /var/lib/lightdm' > /usr/lib/sysusers.d/lightdm.conf
+echo 'u lightdm - "LightDM Daemon" /var/lib/lightdm' > /usr/lib/sysusers.d/lightdm.conf
 systemd-sysusers
 ./configure --prefix=/usr --sysconfdir=/etc --localstatedir=/var --libexecdir=/usr/lib/lightdm --sbindir=/usr/bin --disable-static --disable-tests --with-greeter-user=lightdm --with-greeter-session=lightdm-gtk-greeter
 make
@@ -627,8 +627,8 @@ systemctl enable lightdm
 popd
 rm -rf lightdm-gtk-greeter-2.0.9
 # Firefox.
-[ "$MBS_ARCH" != "x86_64" ] || tar --no-same-owner -xf ../sources/firefox-155.0.tar.xz.1 -C /usr/lib
-[ "$MBS_ARCH" != "aarch64" ] || tar --no-same-owner -xf ../sources/firefox-155.0.tar.xz -C /usr/lib
+[ "$MBS_ARCH" != "x86_64" ] || tar --no-same-owner -xf ../sources/firefox-155.0.1.tar.xz.1 -C /usr/lib
+[ "$MBS_ARCH" != "aarch64" ] || tar --no-same-owner -xf ../sources/firefox-155.0.1.tar.xz -C /usr/lib
 mkdir -p /usr/lib/firefox/distribution
 cat > /usr/lib/firefox/distribution/policies.json << "END"
 {